new func to determine jay response with place activity
The final test determining a jays preferences to a slot, and which slot
it 'checks first' is now a separate function to allow different test
easily.

The new function is place_slot_check, old one is mean_slot_check. It
doesn't behave all that well yet, and outputs and data save need to be
refined.